Set up basic framework for project and towers
I followed a YouTube tutorial series (https://youtube.com/playlist?list=PL_QPQmz5C6WUF-pOQDsbsKbaBZqXj4qSq) which shows how to make a 2D game in Java. The first few videos showed me how to structure this framework, but the video is showing how to make a different type of game, so I had to change many things about the code.

The framework I have here includes a Main.java class which starts the game, a GamePanel.java class which renders the game, and the very basics of towers. I have a superclass Entity.java which is intended to be the parent of both Tower.java and eventually Enemy.java when that is merged in.

Thus far, I have implemented the singleton pattern on GamePanel.java, as we only want one rendering of the game per instance of the application. I have also implemented the Enum pattern for the types of towers so that we can make balance changes to the game all in one place, and there is no chance of creating a tower with the wrong stats.

I also know this should have been broken into smaller commits. I wanted to finish the first few episodes of the tutorial, commit that, and then start making changes, but I didn't want gitHub to keep the files we don't need from the tutorial.